12 hour time format, Mon-Sun datefield option, fix problem with multiple calls to datetimepicker
1. For hour is can show: 12am, 1am,  2am, …. 11pm 12pm and then update
and parse the input field correctly.

Can you make it so the selection dialog for hours:

12am 1am  2am  3am
4am 5am 6am 7am
8am 9am 10am 11am
12pm 1pm 2pm 3pm
4pm 5pm 6pm 7pm
8pm 9pm 10pm 11pm

minutes
12:00 12:05 12:10 12:15
12:20 12:25 12:30 12: 35
12:40 12:45 12:50 12:55

2. Fix for it doesn't always read the input field to add the date to
the picker.  So once you perform and update it will remember the
numbers it had been set to previously.  I need it to always read the
input field and display those as the current date in the date picker.

3. The ability to display the day of the week "Mon, Tues, Wed, Thurs,
Fri, Sat, Sun in the input field and parse it.